Allow auto-merging / upload torrents to the same group + fix error bug
When a group url is specified (or >1 torrent specific url in a later
release) the groupid from the first uploaded torrent is retrieved and
stored and then used in all subsequent uploads in the command. This is
achieved by allowing uploadtorrent() to return the groupid, whether it
be parsed or given to it from a previous run. Fixes #3

Fixed a bug in the error logic, but this needs an overhual