Recent mode with basic size filter (beta)
Adds a new -R / --batchrecent option that allows you to batch upload
torrents from the main torrents.php on JPS. The code does work but the
process is not perfect as unlike with other batchmodes the user will
likely *not* already have the data, consequently when your torrent
client picks up the JPS torrent to start DLing it will be missing so
uploadtorrent() will fail. For recent mode to work you currently need
to run it with it with -d to grab the torrents to start downloading
from JPS, then re-run it without -d but ensure that this time your
JPSTorrents setting in your jps2sm.cfg is set to a different dir,
whilst also ensuring that MediaDirectories is set correctly.

* Parse the main torrent page and send to getbulktorrentids()
* Getting the size of the release and skip if >1Gb, to prevent
  accidentially DLing a huge torrent
* Some refactoring
* Updated some tests

TODO:

* Handle torrent as an in memory object using BytesIO()
* Wait for JPS torrents to download before running uploadtorrent()